requirements.txt
gradio
mcp
openai
python-dotenv
